generic: ag.rs: change (u32, u64) -> (u16, u32).

+++ tests/ag.rs
@@ -6,11 +6,11 @@ use rustad::ag::AG;
 
 #[test]
 fn from_and_to() {
-    let ax  : AG<f64, u32> = AG::from( 3.0f32 );
+    let ax  : AG<f64, u16> = AG::from( 3.0f32 );
     let x  = ax.to_value();
     assert_eq!( x, 3.0f64 );
     //
-    let ax  : AG<f32, u64> = AG::from( 3.0f64 );
+    let ax  : AG<f32, u32> = AG::from( 3.0f64 );
     let x  = ax.to_value();
     assert_eq!( x, 3.0f32 );
 }
